Create IngredientsParse.py
This just parses out the ingredients from the full length recipes, removing stuff from within brackets and deleting extra line feeds and stuff. I didn't write to a csv, because Niels asked me to create a separate file for this, so add the write function where it's commented, or use the function with the path.